+++ Entity.h
@@ -2,16 +2,16 @@
 
 class Entity {
 private:
-	float x;
-	float y;
+	int x;
+	int y;
 	int width;
 	int height;
 
 public:
-	Entity(float newX, float newY, int newWidth, int newHeight);
+	Entity(int newX, int newY, int newWidth, int newHeight);
 	Entity() = default;
-	float getX();
-	float getY();
+	int getX();
+	int getY();
 	int getWidth();
 	int getHeight();
 	void setX(float newX);
